+++ src/Xero.php
@@ -42,8 +42,6 @@ class Xero extends Plugin
 
         self::$plugin = $this;
 
-        $this->_setPluginComponents();
-        $this->_setLogging();
         $this->_registerVariables();
         $this->_registerCraftEventListeners();