+++ src/models/Organisation.php
@@ -105,7 +105,7 @@ class Organisation extends SavableComponent implements OAuthProviderInterface, S
     public function getToken(): ?Token
     {
         if ($this->id) {
-            return Auth::$plugin->getTokens()->getTokenByOwnerReference('commerce-xero', $this->id);
+            return Auth::getInstance()->getTokens()->getTokenByOwnerReference('commerce-xero', $this->id);
         }
 
         return null;